修改和删除收货地址

+++ app/Http/Controllers/UserAddressesController.php
@@ -34,10 +34,12 @@ class UserAddressesController extends Controller
     }
 
     public function edit(UserAddress $user_address ){
+        $this->authorize('own',$user_address);
         return view('user_addresses.create_and_edit',['address'=>$user_address]);
     }
 
     public function update(UserAddress $user_address,UserAddressRequest $request){
+        $this->authorize('own',$user_address);
         $user_address->update($request->only([
             'province',
             'city',
@@ -51,6 +53,7 @@ class UserAddressesController extends Controller
     }
 
     public function destroy(UserAddress $user_address){
+        $this->authorize('own',$user_address);
         $user_address->delete();
         return [];
     }
